Experimented with hashing networks by interactions, and changed gibf to rely only on the genus.
This commit marks a change of direction from IWDB format to WoL format for all networks. This change is made so that networks will be easier to work with, both as excel files and in conversion to pandas dataframes.

In case the WoL format doesn't work, return to this commit.