Updated taxonomy info to implement requests for all 6 web services used by mangal. All services can be queried by a taxonomy id (except for EoL, which shows inconsistent species with mangal), and about half can be queried by name.
The next step is to implement a threadpool to query taxonomy info about all nodes in a network concurrently, then changing the format to a linux/dataframe friendly .csv file (like WoL).
Note that NCBI entrez eutils services limit the rate of requests received from any single IP, so when handling requests keep NCBI as a last resort, and keep it to a small amount of concurrent requests with mutex.