Tested taxonomy information retrieval and completion of kingdom by network structure on networks 1623, 1613 and 950.
Networks 1613 and 1623 showed similar results to network 27: about 90% of the nodes were given a kingdom by taxonomy webservices, and the rest were completed by the network structure with no problem.
Network 950 had an error while getting taxonomy info from web services. Specifically, it appears ITIS jsonservices are down at the moment. As a countermeasure I added a timeout to taxonomy requests, and will deal with it later if the problem persists.